platforms: lm4f: GPIO commit register(CR) should be writable
The commit register needs to be writable in order to be
able to make changes to protected pins PD7, PF0 (NMI)
PC[3:0] (JTAG/SWD).

Before writing to this register the access to it should
be unlocked by writing special value to LOCK register.
After unlocking and writing 1 to corresponding bits in CR
AFSEL, PUR, PDR, DEN bits can be written.